requirements.txt
colorama==0.4.6
playingcards.py==1.1.1
